refactor(shifu): extract the bootstrap engine into a shared crate
Move the pinned-toolchain resolution engine (PATH -> user cache -> download
by exact version, checksum, extract) out of the launcher into a workspace
member the product side can consume, per ADR-0046 stage 1: the runtime
package toolchain (uv, and through it the satellite CPython) must arrive by
the same bootstrap discipline the launcher already implements, shared as a
library rather than reinvented.

crates/bootstrap is std-only with zero external dependencies, same as shifu:
paths (home/cache/PATH probing), fetch (download/sha256/extract), and tool
(the pinned Tool abstraction and bootstrap pipeline). Two deliberate seams
for the trunk consumer: ensure_tool returns Result instead of exiting (each
binary owns its failure policy), and sha256 verification takes an explicit
expectation so a committed manifest can drive it where the launcher uses env
pins. shifu keeps its exact behavior: same messages via the log prefix, same
exit code 127, env-pinned optional checksums, unchanged cache layout.
